+++ misskey-api/src/endpoint/antennas/create.rs
@@ -1,4 +1,4 @@
-#[cfg(feature = "12-10-0")]
+#[cfg(all(feature = "12-10-0", not(feature = "13-7-0")))]
 use crate::model::user_group::UserGroup;
 use crate::model::{
     antenna::{Antenna, AntennaSource},
@@ -21,8 +21,8 @@ pub struct Request {
     pub src: AntennaSource,
     #[builder(default, setter(strip_option))]
     pub user_list_id: Option<Id<UserList>>,
-    #[cfg(feature = "12-10-0")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "12-10-0")))]
+    #[cfg(all(feature = "12-10-0", not(feature = "13-7-0")))]
+    #[cfg_attr(docsrs, doc(cfg(all(feature = "12-10-0", not(feature = "13-7-0")))))]
     #[builder(default, setter(strip_option))]
     pub user_group_id: Option<Id<UserGroup>>,
     #[builder(default, setter(into))]
@@ -64,7 +64,7 @@ mod tests {
                 name: "z0LnEV7NljIUEFFBkjTMW7BN2f6GhfnkbjrNWTqsPikqBzbd02jAvN1axE9h9ZyYCIklKt4WIeeyCNxB31TxJW6hJyHAJVnjTPJC".to_string(),
                 src: AntennaSource::All,
                 user_list_id: None,
-                #[cfg(feature = "12-10-0")]
+                #[cfg(all(feature = "12-10-0", not(feature = "13-7-0")))]
                 user_group_id: None,
                 keywords: Query::from_vec(vec![vec!["hello".to_string(), "awesome".to_string()]]),
                 #[cfg(feature = "12-19-0")]
@@ -88,7 +88,7 @@ mod tests {
                 name: "test".to_string(),
                 src: AntennaSource::Home,
                 user_list_id: None,
-                #[cfg(feature = "12-10-0")]
+                #[cfg(all(feature = "12-10-0", not(feature = "13-7-0")))]
                 user_group_id: None,
                 keywords: Query::from_vec(vec![vec!["hey".to_string()], vec!["wow".to_string()]]),
                 #[cfg(feature = "12-19-0")]
@@ -118,7 +118,7 @@ mod tests {
                 name: "test".to_string(),
                 src: AntennaSource::List,
                 user_list_id: Some(list.id),
-                #[cfg(feature = "12-10-0")]
+                #[cfg(all(feature = "12-10-0", not(feature = "13-7-0")))]
                 user_group_id: None,
                 keywords: Query::from_vec(vec![
                     vec!["kawaii".to_string()],
@@ -139,7 +139,7 @@ mod tests {
     }
 
     #[tokio::test]
-    #[cfg(feature = "12-10-0")]
+    #[cfg(all(feature = "12-10-0", not(feature = "13-7-0")))]
     async fn request_group() {
         use crate::model::{antenna::AntennaSource, query::Query};
 
@@ -184,7 +184,7 @@ mod tests {
                 name: "test".to_string(),
                 src: AntennaSource::Users,
                 user_list_id: None,
-                #[cfg(feature = "12-10-0")]
+                #[cfg(all(feature = "12-10-0", not(feature = "13-7-0")))]
                 user_group_id: None,
                 keywords: Query::from_vec(vec![
                     vec!["annoucement".to_string()],